requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
scikit-learn>=1.3.0
tqdm>=4.65.0
wandb>=0.15.0
betty-ml>=0.2.0
timm>=0.9.0
einops>=0.6.1
imgaug>=0.4.0